+++ src/main/scala/org/podval/tools/cloudrun/CloudRunService.scala
@@ -1,16 +1,12 @@
 package org.podval.tools.cloudrun
 
 import com.google.api.services.run.v1.model.{Configuration, ObjectMeta, Revision, Route, Service}
-import org.slf4j.LoggerFactory
 
-final class CloudRunService(run: CloudRun, val service: Service) {
+final class CloudRunService(run: CloudRun, service: Service) {
 
   private def log: String => Unit = run.log
 
-  private def serviceName: String = service.getMetadata.getName
-
-  // container image name of the first container!
-  def containerImage: String = service.getSpec.getTemplate.getSpec.getContainers.get(0).getImage
+  private def serviceName: String = CloudRun.getServiceName(service)
 
   def deploy(): Service = {
     val previous: Option[Service] = scala.util.Try(get).toOption
@@ -47,17 +43,14 @@ final class CloudRunService(run: CloudRun, val service: Service) {
 
   private def addAnnotations(metadata: ObjectMeta): Unit = {
     val annotations: java.util.Map[String, String] = metadata.getAnnotations
-    annotations.put("run.googleapis.com/client-name", CloudRun.applicationName)
+    annotations.put("run.googleapis.com/client-name"   , CloudRun.applicationName)
     annotations.put("run.googleapis.com/client-version", CloudRun.applicationVersion)
-    annotations.put("client.knative.dev/user-image", containerImage) // TODO why?
+    annotations.put("client.knative.dev/user-image"    , CloudRun.getContainerImage(service)) // TODO why?
   }
 
   def get: Service = run.getService(serviceName)
 
-  def describe(): Unit = log(
-    "Latest Service YAML:\n" +
-    CloudRun.json2yaml(get)
-  )
+  def describe(): Unit = log("Latest Service YAML:\n" + CloudRun.json2yaml(get))
 
   def getConfiguration: Configuration = run.getConfiguration(serviceName)
 
@@ -65,23 +58,5 @@ final class CloudRunService(run: CloudRun, val service: Service) {
 
   def getLatestRevision: Revision = run.getRevision(get.getStatus.getLatestCreatedRevisionName)
 
-  def describeLatestRevision(): Unit = log(
-    "Latest Revision YAML:\n" +
-    CloudRun.json2yaml(getLatestRevision)
-  )
-}
-
-object CloudRunService {
-
-  // manual tests //
-  def main(args: Array[String]): Unit = {
-    val service = new CloudRun(
-      serviceAccountKey = CloudRun.file2string("/home/dub/.gradle/gcloudServiceAccountKey.json"),
-      region = "us-east4",
-      log = LoggerFactory.getLogger(classOf[CloudRunService]).warn
-    ).serviceForYaml("/home/dub/OpenTorah/opentorah.org/collector/service.yaml")
-
-    service.deploy()
-//    println(CloudRun.json2yaml(service.getConfiguration))
-  }
+  def describeLatestRevision(): Unit = log("Latest Revision YAML:\n" + CloudRun.json2yaml(getLatestRevision))
 }