+++ src/main/scala/org/podval/tools/cloudrun/StatusTracker.scala
@@ -1,7 +1,7 @@
 package org.podval.tools.cloudrun
 
 import com.google.api.services.run.v1.model.GoogleCloudRunV1Condition
-import scala.collection.JavaConverters.iterableAsScalaIterableConverter
+import scala.jdk.CollectionConverters.IterableHasAsScala
 
 // see https://github.com/twistedpair/google-cloud-sdk/blob/9d6a1cf6238702560b22944089355eff06b5c216/google-cloud-sdk/lib/googlecloudsdk/api_lib/util/waiter.py
 final class StatusTracker(